+++ x14-2/CommandParser.h
@@ -28,13 +28,13 @@ namespace Windower
 		CommandParser(CommandDispatcher &Dispatcher_in);
 
 		int ParseCommand(const char *pRawCommand_in, WindowerCommand &Command_out,
-						 char **pFeedbackMsg_out, DWORD &FeedbackMsgSize_out);
+						 char **pFeedbackMsg_out, size_t &FeedbackMsgSize_out);
 
 	protected:
-		int Tokenize(const std::string& RawCommand_in, std::string& Command_out, std::list<std::string> &Params_out);
+		size_t Tokenize(const std::string& RawCommand_in, std::string& Command_out, std::list<std::string> &Params_out);
 		int SetFeedback(int Code_in, WindowerCommand &Command_in_out, size_t ParamCount_in,
 						const char *pParamName_in, const char *pParamValue_in,
-						char **pFeedbackMsg_out, DWORD &MsgSize_out);
+						char **pFeedbackMsg_out, size_t &MsgSize_out);
 
 		//! the command dispatcher
 		CommandDispatcher &m_CommandDispatcher;
